Fix copy buttons on Safari and plain-HTTP origins
Extract a shared copyText helper that falls back to execCommand("copy") when
navigator.clipboard is unavailable (non-secure context) or rejects (Safari
outside a trusted gesture). Use it for the assistant id fields, which only
called navigator.clipboard and silently failed; route CopyMessageButton
through it too, dropping its duplicated fallback.
